requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
deepgram-sdk==5.3.0
dnspython==2.8.0
dotenv==0.9.9
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.1
simple-websocket==1.1.0
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
